Add files via upload
The Gaia csv also includes the SDSS data. The MWDD is data from a Montreal databased Dr. Hayden recommended. The Jupyter notebook has the data set up and now we need to build a neural network to find the types of each star. We can use the Montreal data as training and test data.